requirements.txt
absl-py==2.0.0
ipadic==1.0.0
joblib==1.3.2
mecab-python3==1.0.8
nltk==3.8.1
numpy==1.26.1
pandas==2.1.2
torch==2.1.0
tqdm==4.66.1
triton==2.1.0

#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
--no-binary grpcio

accelerate==0.18.0
    # via -r requirements.in
aiohttp==3.8.4
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
datasets==2.9.0
    # via -r requirements.in
dill==0.3.6
    # via
    #   datasets
    #   multiprocess
filelock==3.12.2
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.40.0
    # via matplotlib
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.6.0
    # via
    #   datasets
    #   huggingface-hub
huggingface-hub==0.15.1
    # via
    #   datasets
    #   transformers
idna==3.4
    # via
    #   requests
    #   yarl
importlib-resources==5.12.0
    # via matplotlib
jinja2==3.1.2
    # via torch
kiwisolver==1.4.4
    # via matplotlib
markupsafe==2.1.3
    # via jinja2
matplotlib==3.7.1
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.14
    # via datasets
networkx==3.1
    # via torch
packaging==23.1
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   matplotlib
    #   transformers
pillow==9.5.0
    # via matplotlib
psutil==5.9.5
    # via accelerate
pyarrow==12.0.1
    # via datasets
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2023.3
    # via pandas
pyyaml==6.0
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   transformers
regex==2023.6.3
    # via transformers
requests==2.31.0
    # via
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   responses
    #   transformers
responses==0.18.0
    # via datasets
sentencepiece==0.1.99
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sympy==1.12
    # via torch
tokenizers==0.13.3
    # via transformers
transformers==4.28.1
    # via -r requirements.in
urllib3==2.0.3
    # via
    #   requests
    #   responses
xxhash==3.2.0
    # via datasets
yarl==1.9.2
    # via aiohttp
zipp==3.15.0
    # via importlib-resources